First steps in modernizing Pod::To::HTML
- Use modern file extensions
- Clean up code, use modern features, improve readability
- Add coverage tests

Note: one test in t/120-templates.rakutest needed to be todo'd because
for some reason the more modern Template::Mustache makes that test
fail 9 out of 10 times.  Looks like it's producing the first element
of a hash of about 10 elements, and that produces the wrong thing
most of the time.